refactor(chat): extract ConversationLifecycleHandler from CLVM
PR 4b-1 of the pre-paging cleanup. Pulls the conversation-lifecycle
action-arm cluster out of ConversationListViewModel.kt with no public-API
change. VM drops from 2625 to 2183 lines (-442); the new handler is 561
lines. Combined with PR 4 (which dropped the VM from 3564 to 2625), the
VM is now ~1380 lines lighter than at the start of the cleanup.

The pattern mirrors PR 4 exactly — internal class in the same package,
constructor takes scope/uiState/messagesUiState/services/sendEvent/
dispatch, every arm body is a verbatim move with only mechanical renames
(viewModelScope -> scope, _uiState -> uiState, _messagesUiState ->
messagesUiState, onAction -> dispatch).

What moved
- 26 action arms: Archive*/Unarchive*/ShowArchivedMessagesClicked/
  ArchiveBackClicked, ClearConversation, DeleteConversation,
  ConfirmDeleteConversation, ConfirmLeaveAndDeleteConversation,
  CloseDetailPaneRequestConsumed, AcceptRejoin, DeclineRejoin,
  Introduce*, ConnectIdentities, ConnectToIdentity, AutoConnect,
  OpenConnectionRequestInOwnerConsole, OpenSendConnectionRequestDialog,
  ShowConversationSettings, ShowContactInfo, ShowMessageInfo,
  DismissSheet, TogglePinConversation, BlockUser, ReportContent.
- 6 private helpers: introduceEveryone, runDeleteConversationFlow,
  autoConnect, updateConnectSheetRow, clientExceptionRowState,
  failedOutcomeRowState.

What changed in the VM
- Added the conversationLifecycleHandler property initializer after the
  existing attachmentHandler.
- Replaced 26 arm bodies with single-line delegate calls.
- Deleted the 6 helpers (220 lines removed in the helper region).
- Dropped 27 imports that were only used by the moved code (caught one
  false-positive on chat_introduce_preflight_in_progress, which is still
  used by the post-create preflight collector in the init block, and
  restored it on the next compile).

Verification
- Compiles clean on all 6 targets: homebase-chat:compileKotlinJvm /
  compileAndroidMain / compileKotlinIosSimulatorArm64,
  homebase-core:compileKotlinJvm, androidApp:compileDebugKotlin,
  desktopApp:compileKotlinJvm.
- homebase-chat:jvmTest --rerun-tasks green.
- The runDeleteConversationFlow's `onAction(ClearSelection)` self-dispatch
  becomes `dispatch(ClearSelection)` and routes through the VM's
  onAction. The dispatched action stays in the VM (search/nav cluster),
  so no infinite-loop risk.

Not changed
- Public VM API: uiState, messagesUiState, onAction, eventConsumed,
  dialogClosed, conversationSearchTextState, messageInputTextState,
  messagesSearchTextState — all byte-identical.
- ConversationListScreen.kt, Koin module, nav graph, cross-module
  callers.
- The 23-param VM constructor.

PR 4b-2 (extract MessageActionsHandler) will follow as a separate branch
off the new main once this lands and CI-soaks.
